Reorganize tests and check what files were downloaded
