+++ tests/conftest.py
@@ -5,9 +5,6 @@ from pathlib import Path
 import pytest
 from omegaconf import DictConfig, OmegaConf
 
-from geefetch.cli.omegaconfig import GeefetchConfig, load
-from geefetch.utils.enums import CompositeMethod
-
 TESTS_DIR = Path(__file__).parent
 GEE_PROJECT_ID_ENV_NAME = "GEEFETCH_GEE_PROJECT_ID"
 
@@ -28,41 +25,6 @@ def gee_project_id() -> str:
         case _ as project_id:
             assert isinstance(project_id, str)
             return project_id
-
-
-@pytest.fixture
-def paris_config_path(raw_paris_config: DictConfig, tmp_path: Path, gee_project_id: str) -> Path:
-    raw_paris_config.data_dir = str(tmp_path)
-    raw_paris_config.satellite_default.gee.ee_project_id = gee_project_id
-
-    conf_path = tmp_path / "config.yaml"
-    conf_path.write_text(OmegaConf.to_yaml(raw_paris_config))
-    return conf_path
-
-
-@pytest.fixture
-def paris_timeseriesconfig_path(
-    raw_paris_config: DictConfig, tmp_path: Path, gee_project_id: str
-) -> Path:
-    raw_paris_config.data_dir = str(tmp_path)
-    raw_paris_config.satellite_default.gee.ee_project_id = gee_project_id
-    raw_paris_config.satellite_default.composite_method = CompositeMethod.TIMESERIES
-
-    conf_path = tmp_path / "config.yaml"
-    conf_path.write_text(OmegaConf.to_yaml(raw_paris_config))
-    return conf_path
-
-
-@pytest.fixture
-def paris_config(paris_config_path: Path) -> GeefetchConfig:
-    return load(paris_config_path)
-
-
-@pytest.fixture
-def paris_timeseries_config(paris_config_path: Path) -> GeefetchConfig:
-    config = load(paris_config_path)
-    config.satellite_default.composite_method = CompositeMethod.TIMESERIES
-    return config
 
 
 # ----
